+++ src/main/java/com/pla/individuallife/quotation/presentation/dto/RiderDetailDto.java
@@ -3,7 +3,7 @@ package com.pla.individuallife.quotation.presentation.dto;
 import lombok.Getter;
 import lombok.Setter;
 
-import java.math.BigInteger;
+import java.math.BigDecimal;
 
 /**
  * Created by pradyumna on 05-06-2015.
@@ -13,7 +13,7 @@ import java.math.BigInteger;
 public class RiderDetailDto {
 
     private String coverageId;
-    private BigInteger sumAssured;
+    private BigDecimal sumAssured;
     private Integer coverTerm;
     private Integer waiverOfPremium;
     private String coverageName;
